#1427 Remove Decorator test related stuff. Restore original test package.
