+++ processor/src/test/java/org/mapstruct/ap/test/injectionstrategy/spring/annotateWith/CustomerSpringComponentQualifiedMapper.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.injectionstrategy.spring.annotateWith.qualified;
+package org.mapstruct.ap.test.injectionstrategy.spring.annotateWith;
 
 import org.springframework.stereotype.Component;
