+++ processor/src/test/java/org/mapstruct/ap/test/injectionstrategy/spring/annotateWith/CustomerSpringCustomStereotypeQualifiedMapper.java
@@ -3,17 +3,19 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.injectionstrategy.spring.annotateWith.qualified;
+package org.mapstruct.ap.test.injectionstrategy.spring.annotateWith;
 
 import org.mapstruct.AnnotateWith;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingConstants;
-import org.springframework.stereotype.Service;
 
 /**
  * @author Jose Carlos Campanero Ortiz
  */
-@AnnotateWith( value = Service.class, elements = @AnnotateWith.Element( strings = "AnnotateWithService" ) )
+@AnnotateWith(
+    value = CustomStereotype.class,
+    elements = @AnnotateWith.Element( strings = "AnnotateWithCustomStereotype" )
+)
 @Mapper( componentModel = MappingConstants.ComponentModel.SPRING )
-public interface CustomerSpringServiceQualifiedMapper {
+public interface CustomerSpringCustomStereotypeQualifiedMapper {
 }
