+++ processor/src/test/java/org/mapstruct/ap/test/injectionstrategy/spring/annotateWith/CustomerSpringRepositoryQualifiedMapper.java
@@ -3,19 +3,17 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.injectionstrategy.spring.annotateWith.qualified;
+package org.mapstruct.ap.test.injectionstrategy.spring.annotateWith;
 
 import org.mapstruct.AnnotateWith;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingConstants;
+import org.springframework.stereotype.Repository;
 
 /**
  * @author Jose Carlos Campanero Ortiz
  */
-@AnnotateWith(
-    value = CustomStereotype.class,
-    elements = @AnnotateWith.Element( strings = "AnnotateWithCustomStereotype" )
-)
+@AnnotateWith( value = Repository.class, elements = @AnnotateWith.Element( strings = "AnnotateWithRepository" ) )
 @Mapper( componentModel = MappingConstants.ComponentModel.SPRING )
-public interface CustomerSpringCustomStereotypeQualifiedMapper {
+public interface CustomerSpringRepositoryQualifiedMapper {
 }
