--- processor/src/test/java/org/mapstruct/ap/test/injectionstrategy/spring/annotateWith/decorator/GenderSpringDefaultMapper.java
@@ -1,26 +1,0 @@
-/*
- * Copyright MapStruct Authors.
- *
- * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
- */
-package org.mapstruct.ap.test.injectionstrategy.spring.annotateWith.decorator;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.MappingConstants;
-import org.mapstruct.ValueMapping;
-import org.mapstruct.ValueMappings;
-import org.mapstruct.ap.test.injectionstrategy.shared.Gender;
-import org.mapstruct.ap.test.injectionstrategy.shared.GenderDto;
-
-/**
- * @author Filip Hrisafov
- */
-@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
-public interface GenderSpringDefaultMapper {
-
-    @ValueMappings({
-            @ValueMapping(source = "MALE", target = "M"),
-            @ValueMapping(source = "FEMALE", target = "F")
-    })
-    GenderDto mapToDto(Gender gender);
-}
